requirements.txt
python-dotenv==1.2.2
pandas==3.0.2
matplotlib==3.10.8
biopython==1.87
seaborn==0.13.2
h5py==3.16.0
torchmetrics==1.9.0
fastprogress==1.0.3
scipy==1.17.1
scikit-learn==1.8.0
gdown==3.4.6
torch